+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Note that since we don't clearly distinguish between a public and private interf
 
 - Fix marking pass not working with ``transparentBackground``
 - Fix pdbe xray maps url not https
+- Fix entity-id color theme broken for non-IHM models
 
 ## [v3.0.0] - 2022-01-23