+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix marking pass not working with ``transparentBackground``
 - Fix pdbe xray maps url not https
 - Fix entity-id color theme broken for non-IHM models
+- Improve/fix marking of ``InteractionsInterUnitVisual`` (mark when all contact-feature members are given)
 
 ## [v3.0.0] - 2022-01-23